expanded system to work with up to 4 splits. Need to redo terrain system to handle this way better. Fixed some major precision issues in CTimer Added color parameter to materials, custom parameters to come in the future Improved CRenderTexture to handle more possible cases Added way to compile shader from string to CShader Major improvements to material system, added surface shaders and ability to include other materials

+++ JetEngine/Util/CTimer.cpp
@@ -110,9 +110,9 @@ void CTimer::Update()
     m_numFrames++;
     if ( m_currentTime - m_lastFPSUpdate >= m_FPSUpdateInterval )
     {
-        float currentTime = (float)m_currentTime / (float)m_ticksPerSecond;
-        float lastTime = (float)m_lastFPSUpdate / (float)m_ticksPerSecond;
-        m_fps = (float)m_numFrames / (currentTime - lastTime);
+        double currentTime = (double)m_currentTime / (double)m_ticksPerSecond;
+		double lastTime = (double)m_lastFPSUpdate / (double)m_ticksPerSecond;
+		m_fps = (double)m_numFrames / (currentTime - lastTime);
 
         m_lastFPSUpdate = m_currentTime;
         m_numFrames = 0;
